Fix default download URL path
The download URL path for older versions seems to have changed. So, the
download of artifacts results in a 404 for cases where a specific
release version is being defined (see tasks 'Compose ... download url)'.
The proposed adjustments aims to fix this issue.

Note that if a release version is not defined explicitly, the proposed
change still works, because the latest versions are placed in the archive
folder, too.