Remove the role's group-awareness
Up until now the role requires a certain inventory structure. This change
set aims to lift that restriction and thus allow a more flexible usage.
It is now possible to solely let the play be in charge of determine the hosts,
either by setting the desired 'hosts' with a pattern [1] or by overriding
'minio_hosts' in the 'vars' attribute of the play.

## Notable changes

* now, by default the role is applied to each host that the play's 'hosts'
  attribute refers to or implies (through group names); it's thus determined
  outside of the role, which is the preferred place
* this way, the role inside does not rely on a default group value or any
  group at all
* the Molecule scenario named 'layouts' is being used to test this behaviour,
  whereas the 'client' scenario explicitly uses the group assignment

## Migration path

A change is only required if the play that uses this role somehow
refers to an overall set of 'hosts' that contains nodes which are
not supposed to get this role assigned to, e.g.: the 'hosts' value
shows 'all' and the inventory holds another distinct group of
$database hosts.

[1] https://docs.ansible.com/ansible/latest/user_guide/intro_patterns.html